pyproject.toml
[tool.poetry]
name = "mmcifix"
version = "0.1.0"
description = ""
authors = ["Michael Milton <michael.r.milton@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6.0,<4"
biopython = "^1.79"
more-itertools = "^8.12.0"
click = "^7.0.0"
inflection = "^0.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mmcifix = 'mmcifix:main'
